[ThoiND][Order] 60% payment order 19/10/2023 08:27AM

+++ src/main/java/com/example/c4zone/model/order/Cart.java
@@ -17,7 +17,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idCart;
     @Column(name ="quantity_product_order")
-    private Long quantityProductOrder;
+    private Integer quantityProductOrder;
 
 
     @ManyToOne
